fix: resolve mypy type errors in cache, eval, and guardrails modules
- prompt_injection.py: Fix LLM usage (use chat() instead of generate())
- prompt_injection.py: Add Literal type annotation for severity
- topic_filter.py: Fix Embedder to Embeddings import
- topic_filter.py: Add Literal type annotation for severity
- topic_filter.py: Cast _cosine_similarity fallback to float
- evaluators.py: Fix LLM constructor and use achat() instead of ainvoke()
- rag.py: Ensure str() cast on generator return values

+++ src/ai_infra/cache/base.py
@@ -300,7 +300,7 @@ class CacheBackend(ABC):
         if norm_a == 0 or norm_b == 0:
             return 0.0
 
-        return dot_product / (norm_a * norm_b)
+        return float(dot_product / (norm_a * norm_b))
 
     @staticmethod
     def hash_key(key: str) -> str:
